BOL-375 refactor: Repository Impl 파일들 서버 맞춰서 리팩토링
