+++ data/src/main/java/com/yapp/bol/data/repository/AuthRepositoryImpl.kt
@@ -1,6 +1,8 @@
 package com.yapp.bol.data.repository
 
 import com.yapp.bol.data.datasource.auth.AuthDataSource
+import com.yapp.bol.data.mapper.AuthMapper.toDomain
+import com.yapp.bol.domain.model.LoginItem
 import com.yapp.bol.domain.repository.AuthRepository
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
@@ -29,5 +31,9 @@ class AuthRepositoryImpl @Inject constructor(
 
     override suspend fun deleteRefreshToken() {
         dataSource.deleteRefreshToken()
+    }
+
+    override suspend fun login(type: String, token: String): LoginItem? {
+        return dataSource.login(type, token).toDomain()
     }
 }
